Renaming preview files and imports.
