+++ contentcuration/contentcuration/static/js/edit_channel/preview/views/FilePreview.js
