+++ contentcuration/contentcuration/static/js/edit_channel/preview/views/index.js
@@ -1,4 +1,3 @@
 // maintaining old API
-export { default as PreviewModalView } from './previewModalView.js';
-export { default as PreviewView } from './previewManagerView.js';
-export { default as ItemPreviewView } from './previewView.js';
+export { default as NodePreview } from './NodePreview.js';
+export { default as FilePreview } from './FilePreview.js';
